pointers, Arrays and strings

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -8,11 +8,12 @@
  * @str: The string to get the length of.
  * Return: The lenght of @str
  */
-int _strlen(const char *s)
+size_t _strlen(const char *s)
 {
 	size_t length = 0
 
 	while (*str++)
 		length++;
+
 	return (length);
 }
